update build scripts

+++ scripts/adduser_lfs-4.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-set -e
 if [ `id -u` != 0 ];then
         echo Permission delay, Please run as root!
         exit
@@ -9,35 +8,17 @@ fi
 export LFS=/mnt/lfs
 
 #4.2. 在 LFS 文件系统中创建有限目录布局
-mkdir -pv $LFS/{bin,etc,lib,sbin,usr,var}
-case $(uname -m) in
-  x86_64) mkdir -pv $LFS/lib64 ;;
-esac
-
-mkdir -pv $LFS/tools
+. build/cmd/chapter04/creatingminlayout.cmd
 
 #4.3. 添加 LFS 用户
-set +e
-userdel lfs
-set -e
-rm /home/lfs -rf
-groupadd lfs
-useradd -s /bin/bash -g lfs -m -k /dev/null lfs
-echo "Please set the new account lfs's password:"
+#echo "Please set the new account lfs's password:"
 #passwd lfs
-echo lfs:1 | chpasswd
-
-chown -R lfs $LFS
-chgrp -R lfs $LFS
-case $(uname -m) in
-  x86_64) chown -v lfs $LFS/lib64 ;;
-esac
-
-#copy object compile file to lfs account
-cp -r ../haoos-lfs /home/lfs/
-
-#change the owner and group of /home/lfs from root to lfs.
-chown -v -R lfs /home/lfs
-chgrp -v -R lfs /home/lfs
-
-su - lfs
+#echo lfs:1 | chpasswd
+getent group | grep lfs
+if [ $? -eq 0 ];then
+	userdel lfs
+	groupdel lfs
+	rm /home/lfs -rf
+fi
+sed -i 's/passwd lfs/echo lfs:1 | chpasswd/g' build/cmd/chapter04/addinguser.cmd
+. build/cmd/chapter04/addinguser.cmd
